BROKEN: Work in Progress: adding builtin functions to tests
