+++ tests/ReflectionParameterTest.php
@@ -5,8 +5,13 @@ use Go\ParserReflection\Stub\Foo;
 use Go\ParserReflection\Stub\SubFoo;
 use TestParametersForRootNsClass;
 
-class ReflectionParameterTest extends \PHPUnit_Framework_TestCase
+class ReflectionParameterTest extends TestCaseBase
 {
+    /**
+     * @var string
+     */
+    protected $lastFileSetUp;
+
     /**
      * @var ReflectionFile
      */
@@ -18,11 +23,10 @@ class ReflectionParameterTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @dataProvider fileProvider
+     * @dataProvider functionProvider
      */
-    public function testGeneralInfoGetters($fileName)
+    public function testGeneralInfoGetters($funcName, $fileName)
     {
-        $this->setUpFile($fileName);
         $allNameGetters = [
             'isArray', 'isCallable', 'isOptional', 'isPassedByReference', 'isDefaultValueAvailable',
             'getPosition', 'canBePassedByValue', 'allowsNull', 'getDefaultValue', 'getDefaultValueConstantName',
@@ -38,28 +42,34 @@ class ReflectionParameterTest extends \PHPUnit_Framework_TestCase
             $allNameGetters[] = 'hasType';
         }
 
-        foreach ($this->parsedRefFile->getFileNamespaces() as $fileNamespace) {
-            foreach ($fileNamespace->getFunctions() as $refFunction) {
-                $functionName = $refFunction->getName();
-                foreach ($refFunction->getParameters() as $refParameter) {
-                    $parameterName        = $refParameter->getName();
-                    $originalRefParameter = new \ReflectionParameter($functionName, $parameterName);
-                    foreach ($allNameGetters as $getterName) {
+        if ($fileName) {
+            $this->setUpFile($fileName);
+            $fileNamespace = $this->parsedRefFile->getFileNamespace(
+                $this->getNamespaceFromName($funcName));
+            $refFunction   = $fileNamespace->getFunction($funcName);
+        } else {
+            $this->lastFileSetUp = null;
+            $this->parsedRefFile = null;
+            $refFunction = new ReflectionFunction($funcName);
+        }
+        $functionName  = $refFunction->getName();
+        foreach ($refFunction->getParameters() as $refParameter) {
+            $parameterName        = $refParameter->getName();
+            $originalRefParameter = new \ReflectionParameter($functionName, $parameterName);
+            foreach ($allNameGetters as $getterName) {
 
-                        // skip some methods if there is no default value
-                        $isDefaultValueAvailable = $originalRefParameter->isDefaultValueAvailable();
-                        if (isset($onlyWithDefaultValues[$getterName]) && !$isDefaultValueAvailable) {
-                            continue;
-                        }
-                        $expectedValue = $originalRefParameter->$getterName();
-                        $actualValue   = $refParameter->$getterName();
-                        $this->assertSame(
-                            $expectedValue,
-                            $actualValue,
-                            "{$getterName}() for parameter {$functionName}:{$parameterName} should be equal"
-                        );
-                    }
+                // skip some methods if there is no default value
+                $isDefaultValueAvailable = $originalRefParameter->isDefaultValueAvailable();
+                if (isset($onlyWithDefaultValues[$getterName]) && !$isDefaultValueAvailable) {
+                    continue;
                 }
+                $expectedValue = $originalRefParameter->$getterName();
+                $actualValue   = $refParameter->$getterName();
+                $this->assertSame(
+                    $expectedValue,
+                    $actualValue,
+                    "{$getterName}() for parameter {$functionName}:{$parameterName} should be equal"
+                );
             }
         }
     }
@@ -88,7 +98,7 @@ class ReflectionParameterTest extends \PHPUnit_Framework_TestCase
      *
      * @return array
      */
-    public function getFunctionsToAnalyze()
+    public function functionProvider()
     {
         // Random selection of built in functions.
         $builtInFunctions = ['preg_match', 'date', 'create_function'];
@@ -96,7 +106,7 @@ class ReflectionParameterTest extends \PHPUnit_Framework_TestCase
         foreach ($builtInFunctions as $functionsName) {
             $functions[$functionsName] = ['function' => $functionsName, 'fileName'  => null];
         }
-        $files = $this->getFilesToAnalyze();
+        $files = $this->fileProvider();
         foreach ($files as $filenameArgList) {
             $argKeys = array_keys($filenameArgList);
             $fileName = $filenameArgList[$argKeys[0]];
@@ -334,11 +344,14 @@ class ReflectionParameterTest extends \PHPUnit_Framework_TestCase
     private function setUpFile($fileName)
     {
         $fileName = stream_resolve_include_path($fileName);
-        $fileNode = ReflectionEngine::parseFile($fileName);
+        if ($this->lastFileSetUp !== $fileName) {
+            $fileNode = ReflectionEngine::parseFile($fileName);
 
-        $reflectionFile = new ReflectionFile($fileName, $fileNode);
-        $this->parsedRefFile = $reflectionFile;
+            $reflectionFile = new ReflectionFile($fileName, $fileNode);
+            $this->parsedRefFile = $reflectionFile;
 
-        include_once $fileName;
+            include_once $fileName;
+            $this->lastFileSetUp = $fileName;
+        }
     }
 }
